+++ models/jwt_claims.go
@@ -3,7 +3,7 @@ package models
 import "github.com/golang-jwt/jwt/v4"
 
 type JwtCustomClaims struct {
-	UserId     uint   `json:"user_id"`
+	UserId     int64  `json:"user_id"`
 	UserSecret string `json:"user_secret"`
 	jwt.RegisteredClaims
 }
