+++ models/user.go
@@ -1,5 +1,5 @@
 package models
 
 type User struct {
-	Id uint `json:"user_id"`
+	Id int64 `json:"user_id"`
 }
